pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gge-utility-bot"
version = "1.0.0"
description = "A discord bot with a variety of functions for GGE"

dependencies = [
    "discord.py==2.6.3", 
    "websockets==14.1", 
    "Quart==0.20.0", 
    "cryptography==43.0.1", 
    "pillow==12.0.0", 
    "numpy==2.3.4", 
    "pydantic==2.12.4", 
    "typing_extensions==4.15.0",
    "pymongo==4.15.5",
]

[project.optional-dependencies]
test = [
    "pytest==8.4.2", 
    "pytest-asyncio==1.2.0", 
    "pytest-mock==3.15.1",
]